These are the doctor view fxml and controller files
there are 4 doctors views namely:
docView which is the dashboard
doctorMessages is the message view for doctors
doctorpatientRecords is the patient record for doctors to view
doctorPrescription is the order prescription page. 

I have not commented the code yet i will push the comments soon for easier understanding. All the fxml files have seperate controller files.